pyproject.toml
[project]
name = "portovan"
version = "0.1.0"
description = "Interactive Leaflet map for a van trip in Portugal"
requires-python = ">=3.10"
dependencies = [
    "folium",
    "pandas",
    "playwright",
    "python-dotenv",
    "requests",
]